pyproject.toml
[project]
name = "plan-mcp-ag-server"
version = "0.1.0"
description = "Oracle EPM Planning Agentic MCP Server using Google ADK"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Planning Team"}
]

dependencies = [
    # Core ADK
    "google-adk>=1.0.0",

    # MCP Protocol (for Claude Desktop)
    "mcp>=1.0.0",

    # Configuration
    "pydantic-settings>=2.0.0",

    # HTTP client (async)
    "httpx>=0.27.0",

    # Database (PostgreSQL for sessions + feedback)
    "sqlalchemy>=2.0.0",
    "psycopg[binary]>=3.1.0",

    # Document generation
    "python-docx>=1.1.0",

    # Web server
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",

    # Reinforcement Learning
    "numpy>=1.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
]
dashboard = [
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
    "pandas>=2.0.0",
]

[project.scripts]
plan-mcp = "cli.mcp_server:main"
plan-web = "web.server:main"
plan-cli = "cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["planning_agent", "web", "cli"]


